Set up defining/editing/displaying press kit page
Defined the content types & parts for editing & displaying a press kit
page. These are generated when the module is enabled via a recipe
migration. The definitions uses named parts to improve breaking up the
editor with titles.

Unfortunately because of using named parts the majority of the display
is hardcoded within the content shape. Was unable to get the
`placement.json` to hook up with defining the position for named parts.