Add option to display as per presskit()
Added option to "Display as presskit()", which will render the press kit
using the same display as [presskit()](https://dopresskit.com). The way
it works is when the `UsePressKitDisplay` is `true`, the driver will
redirect the user to a controller route that handles rendering a page
with no theme assets and uses the assets from presskit() and uses the
same HTML structure.